Rename DatabaseConfigurations::RootConfig to BaseConfig
and extract BaseConfig and TenantConfig into separate files.
