+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # `activerecord-tenanted` Changelog
 
+## next / unreleased
+
+- Rename `ActiveRecord::Tenanted::DatabaseConfigurations::RootConfig` to `BaseConfig`.
+
+
 ## 0.3.0 / 2025-09-09
 
 ### Improved
