+++ GUIDE.md
@@ -390,10 +390,10 @@ Documentation outline:
 
 TODO:
 
-- implement `AR::Tenanted::DatabaseConfigurations::RootConfig`
-  - [x] create the specialized RootConfig for `tenanted: true` databases
-  - [x] RootConfig disables database tasks initially
-  - [x] RootConfig raises if a connection is attempted
+- implement `AR::Tenanted::DatabaseConfigurations::BaseConfig`
+  - [x] create the specialized BaseConfig for `tenanted: true` databases
+  - [x] BaseConfig disables database tasks initially
+  - [x] BaseConfig raises if a connection is attempted
   - [x] `#database_path_for(tenant_name)`
   - [x] `#tenants` returns all the tenants on disk (for iteration)
   - [x] raise an exception if tenant name contains a path separator
